consolidated all functions into one logic function
